kudumu test ettim. orijinalprintf çıktıları  ile arasında orantısızlıklar mevcut. sorunları gidermek için gerekeni yapmaya başlıyorum.

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: bayram-seven <bayram-seven@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/31 22:24:31 by bayram-seve       #+#    #+#             */
-/*   Updated: 2026/04/02 05:02:17 by bayram-seve      ###   ########.fr       */
+/*   Updated: 2026/04/02 05:21:11 by bayram-seve      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	ft_print_char(int chr);
 int ft_print_str(char *str);
 int ft_print_nbr(int n);
 int ft_print_unbr(unsigned int n);
-int ft_print_hexnum( unsigned int n, char format );
+int ft_print_hexnum( unsigned long long n, char format );
 int	ft_print_ptr(unsigned long long ptr);
 
 int ft_check_format(va_list argm, const char type );
